refactor(api): in team, use admin-member repo from team
