+++ api/tests/team/integration/infrastructure/repositories/admin-member.repository.test.js
@@ -2,8 +2,8 @@ import sinon from 'sinon';
 
 import { PIX_ADMIN } from '../../../../../src/authorization/domain/constants.js';
 import { AdminMemberError } from '../../../../../src/authorization/domain/errors.js';
-import { adminMemberRepository } from '../../../../../src/shared/infrastructure/repositories/admin-member.repository.js';
 import { AdminMember } from '../../../../../src/team/domain/models/AdminMember.js';
+import { adminMemberRepository } from '../../../../../src/team/infrastructure/repositories/admin-member.repository.js';
 import { expect } from '../../../../test-helper.js';
 import { databaseBuilder, knex } from '../../../../tooling/databases.js';
 import { catchErr } from '../../../../tooling/test-utils/error.js';
